refactor: test_evaluate_candidate_node.pyを機能別に3ファイルに分割
- test_evaluate_candidate_node_main.py: evaluate_candidate_node関数のテスト
- test_comment_evaluator.py: CommentEvaluatorクラスのテスト
- test_evaluation_criteria.py: 評価基準データクラスのテスト

また、evaluate_candidate_node.pyの_restore_weather_data関数を修正し、
WeatherForecastの必須パラメータ（weather_condition, wind_direction_degrees）を
適切に処理するように更新

テスト数と成功/失敗の状態は変更なし（15テスト、6失敗）
